requirements.txt
annotated-types==0.7.0
awsebcli==3.22.1
blinker==1.9.0
boto3==1.40.52
botocore==1.40.74
cachetools==6.2.1
cement==2.10.14
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
Flask==3.0.3
Flask-Cors==4.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
packaging==24.2
pathspec==0.10.1
pillow==12.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
PyPDF2==3.0.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==311
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
s3transfer==0.14.0
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
termcolor==2.5.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==1.26.20
wcwidth==0.2.14
Werkzeug==3.1.3
